requirements.txt
# ThermaSched Python dependencies
# Install with: pip install -r requirements.txt

# Core scientific computing
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
pandas>=2.0.0

# Discrete-event simulation (scalability projections)
simpy>=4.1.1

# Hardware interface
pyserial>=3.5

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Data analysis and visualization
seaborn>=0.12.0
jupyter>=1.0.0
ipykernel>=6.0.0

# Curve fitting and signal processing
scikit-learn>=1.3.0

# Report generation
jinja2>=3.1.0
tabulate>=0.9.0
